Fix issue #1 (Buffer lag)
Based on emperical evidence, the i2s_channel_write does not flush the
most recent buffer to the hardware. A workaround to this was added,
by sending the "last" buffer a second time before blocking to wait for
another pixel update. This does not affect performance when the pixel
buffer is updated continually.

Some fixes to the nexopixel_Deinit process were added as well.

Unit test app updated to demonstrate changes in this commit.